Bug 1256295 - Reorganize releasetasks to handle multiple products
