+++ releasetasks/templates/firefox/bouncer.yml.tmpl
