+++ releasetasks/templates/firefox/checksums.yml.tmpl
