+++ releasetasks/templates/firefox/l10n.yml.tmpl
