+++ releasetasks/templates/firefox/release_graph.yml.tmpl
