+++ releasetasks/templates/firefox/source.yml.tmpl
