+++ releasetasks/templates/firefox/updates.yml.tmpl
