+++ releasetasks/templates/firefox/uptake_monitoring.yml.tmpl
