+++ releasetasks/templates/firefox/version_bump.yml.tmpl
