+++ releasetasks/test/firefox/test_bouncer_submission.py
@@ -1,14 +1,14 @@
 import unittest
 
-from releasetasks.test import make_task_graph, PVT_KEY_FILE, \
-    do_common_assertions, get_task_by_name
+from releasetasks.test.firefox import make_task_graph, do_common_assertions, \
+    get_task_by_name
+from releasetasks.test import PVT_KEY_FILE
 
 
-class TestVersionBump(unittest.TestCase):
+class TestBouncerSubmission(unittest.TestCase):
     maxDiff = 30000
     graph = None
     task = None
-    human_task = None
     payload = None
 
     def setUp(self):
@@ -18,7 +18,6 @@ class TestVersionBump(unittest.TestCase):
             appVersion="42.0",
             buildNumber=3,
             source_enabled=False,
-            checksums_enabled=False,
             en_US_config={
                 "platforms": {
                     "macosx64": {},
@@ -44,9 +43,10 @@ class TestVersionBump(unittest.TestCase):
             branch="foo",
             updates_enabled=False,
             bouncer_enabled=True,
+            checksums_enabled=False,
             push_to_candidates_enabled=False,
             push_to_releases_enabled=False,
-            postrelease_version_bump_enabled=True,
+            postrelease_version_bump_enabled=False,
             signing_class="release-signing",
             release_channels=["foo"],
             balrog_api_root="https://balrog.real/api",
@@ -54,41 +54,34 @@ class TestVersionBump(unittest.TestCase):
             signing_pvt_key=PVT_KEY_FILE,
             build_tools_repo_path='build/tools',
         )
-        self.task = get_task_by_name(
-            self.graph, "release-foo-firefox_version_bump")
-        self.human_task = get_task_by_name(
-            self.graph, "release-foo-firefox_version_bump_human_decision")
+        self.task = get_task_by_name(self.graph,
+                                     "release-foo_firefox_bncr_sub")
         self.payload = self.task["task"]["payload"]
 
     def test_common_assertions(self):
         do_common_assertions(self.graph)
 
     def test_provisioner(self):
-        self.assertEqual(self.task["task"]["provisionerId"],
-                         "buildbot-bridge")
-
-    def test_human_provisioner(self):
-        self.assertEqual(self.human_task["task"]["provisionerId"],
-                         "null-provisioner")
+        self.assertEqual(self.task["task"]["provisionerId"], "buildbot-bridge")
 
     def test_worker_type(self):
         self.assertEqual(self.task["task"]["workerType"], "buildbot-bridge")
 
-    def test_human_worker_type(self):
-        self.assertEqual(self.human_task["task"]["workerType"],
-                         "human-decision")
+    def test_scopes_present(self):
+        self.assertFalse("scopes" in self.task)
 
-    def test_next_version(self):
-        self.assertEqual(self.payload["properties"]["next_version"], "42.0b3")
+    def test_partials(self):
+        self.assertEqual(self.payload["properties"]["partial_versions"],
+                         "37.0build2, 38.0build1")
+
+    def test_build_number(self):
+        self.assertEqual(self.payload["properties"]["build_number"], 3)
 
     def test_graph_scopes(self):
         expected_graph_scopes = set([
             "queue:task-priority:high",
         ])
         self.assertTrue(expected_graph_scopes.issubset(self.graph["scopes"]))
-
-    def test_requires(self):
-        self.assertIn(self.human_task["taskId"], self.task["requires"])
 
     def test_repo_path(self):
         self.assertEqual(self.payload["properties"]["repo_path"],
