+++ releasetasks/test/firefox/test_l10n.py
@@ -1,8 +1,8 @@
 import unittest
 
-from releasetasks.test import make_task_graph, PVT_KEY_FILE, \
-    get_task_by_name, \
+from releasetasks.test.firefox import make_task_graph, get_task_by_name, \
     do_common_assertions
+from releasetasks.test import PVT_KEY_FILE
 
 
 class TestL10NSingleChunk(unittest.TestCase):
